Wystawienie resta ze swoimi rezerwacjami oraz anulacja rezerwacji.
