+++ src/main/java/com/inz/airline/controller/JourneyController.java
@@ -33,13 +33,12 @@ public class JourneyController {
 
     @PostMapping("/journey")
     ResponseEntity<List<Journey>> getJourney(@RequestBody SearchFlightDto searchFlightDto){
-        System.out.println("data sercha "+searchFlightDto.getDataStartSearch());
         return new ResponseEntity<>(journeyService.getJourney(searchFlightDto), HttpStatus.ACCEPTED);
     }
 
 
 
-    @GetMapping("getAll")
+    @GetMapping("getAllJourney")
     ResponseEntity<List<Journey>> getjourney(){
         return new ResponseEntity<>((List<Journey>) journeyRepository.findAll(), HttpStatus.ACCEPTED);
 
@@ -48,6 +47,23 @@ public class JourneyController {
     ResponseEntity<List<Booking>> getBookings(){
         return new ResponseEntity<>((List<Booking>) bookingRepository.findAll(), HttpStatus.ACCEPTED);
 
+    }
+    @GetMapping("reservations")
+    ResponseEntity<List<Booking>> getMyReservations(){
+        User principal = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+
+        return new ResponseEntity<>(bookingService.getMyReservations(principal.getUsername()), HttpStatus.ACCEPTED);
+
+    }
+
+    @PostMapping("/booking")
+    ResponseEntity<Booking> bookFlights(@RequestBody BookingDto bookingDto){
+        User principal = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+
+        System.out.println(principal);
+        bookingDto.setOwner(principal.getUsername());
+
+        return new ResponseEntity<>(bookingService.addBooking(bookingDto), HttpStatus.ACCEPTED);
     }
 
     @PostMapping("/cancelBooking")
